Split out tokens.py file. Bug fixes and tests
Crucially, I haven't written any tests for either `tokenize` itself
or `_consume_next_symbol`, `_consume_next_string` or
`_consume_next_literal`. At this point a metric boatload of unit tests
is in order since to debug the parser we'll first have to be able to
convince ourselves that the lexer is correct.